moved the JsonSerializationContext/JsonDeserializationContext fields to where they are used. Also moved TreeTypeAdapter to internal.bind package for potential use in JsonAdapterAnnotationFactory.
